requirements.txt
rpyc
opencv-python
numpy
matplotlib
torch
torchvision
timm
pillow
ipywidgets
ipython
tqdm
wandb